backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.5

# LangChain + FAISS
langchain==0.2.16
langchain-community==0.2.16
faiss-cpu==1.8.0.post1

# OpenAI
openai==1.43.0

# Text loaders/parsers
unstructured==0.15.12
pdfminer.six==20231228

mammoth==1.8.0
